requirements.txt
pycryptodome
cryptography
argparse